18-01-2021

+++ admin/gestion_avis.php
@@ -5,6 +5,15 @@ if (!internauteEstConnecteEtEstAdmin())
   header('location:../pages/connexion.php');
   exit();
 }
+
+//suppression d'un avis
+if(isset($_GET['action']) && $_GET['action'] == 'delete')
+{   
+   $statement = $pdo->prepare("DELETE FROM avis WHERE id_avis = ?");
+   $resultat = $statement->execute(array($_GET['id']));
+   $message = '<div class="alert alert-success" role="alert">Suppression de l\'avis '.$_GET['id'].'</div>';
+  }
+
 ?>
 <?php include_once("../inc/haut.inc.php");?>
 <?php include_once('../inc/menu.inc.php');?>
@@ -12,21 +21,9 @@ if (!internauteEstConnecteEtEstAdmin())
 <div id="gestion_avis" class="conteneur py-5">
     <h1>Gestion des avis</h1>
     <div class="text-center mb-3">
+<?php echo $message ?>
 
-
-  <?php
-
-//suppression d'un avis
-if(isset($_GET['action']) && $_GET['action'] == 'delete')
-{   
-   $statement = $pdo->prepare("DELETE FROM avis WHERE id_avis = ?");
-   $resultat = $statement->execute(array($_GET['id']));
-   echo '<div class="alert alert-success" role="alert">Suppression de l\'avis '.$_GET['id'].' effectué</div><br><a href="/admin/gestion_avis.php">Retour sur la page des avis</a>' ;
-      die;
-}
-
-//Affichage des salles
-?>
+<!-- Affichage des salles -->
     <div class="row">
  
         <table class="table table-striped text-center">
@@ -47,6 +44,7 @@ if(isset($_GET['action']) && $_GET['action'] == 'delete')
 $resultat = $pdo->query('SELECT * FROM avis');
 while ($datas = $resultat->fetch(PDO::FETCH_ASSOC)) 
 {
+    $datas['commentaire']=stripslashes($datas['commentaire']);
     echo '<tr>';
     echo '<td>'.$datas['id_avis'].'</td>';
     echo '<td>'.$datas['id_membre'].'</td>';
@@ -54,7 +52,7 @@ while ($datas = $resultat->fetch(PDO::FETCH_ASSOC))
     echo '<td>'.$datas['commentaire'].'</td>';
     echo '<td>'.$datas['note'].'</td>';
     echo '<td>'.$datas['date'].'</td>';
-    echo '<td><a href="?action=delete&id='.$datas['id_avis'].'"><i class="fa fa-trash fa-2x"></i></a></td>';
+    echo '<td><a href="?action=delete&id='.$datas['id_avis'].'" Onclick="'."return(confirm('Êtes-vous sûr de vouloir supprimer l'avis ?'))".'"><i class="fa fa-trash fa-2x"></i></a></td>';
     echo '</tr>';
 }
 ?>
